Extract last-mod check to more generic package
This allows us to share the last modification logic for any directory,
not just a schema.Issue
